+++ src/data/skill-check/action/demoralize.ts
@@ -10,7 +10,8 @@ export const actions: AssistantAction[] = [
             if (!message.target?.actor) return;
 
             const frightened = message.target.actor.getCondition("frightened");
-            if (frightened?.value ?? 0 < 2) {
+            const value = frightened?.value ?? 0;
+            if (value < 2) {
                 await game.assistant.socket.increaseCondition(message.target.actor, "frightened", { max: 2, value: 2 });
             }
         },
@@ -23,7 +24,8 @@ export const actions: AssistantAction[] = [
             if (!message.target?.actor) return;
 
             const frightened = message.target.actor.getCondition("frightened");
-            if (frightened?.value ?? 0 < 1) {
+            const value = frightened?.value ?? 0;
+            if (value < 1) {
                 await game.assistant.socket.increaseCondition(message.target.actor, "frightened", { max: 1, value: 1 });
             }
         },