+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
   - This is reliant on whoever is rolling to have 3D dice enabled, if they have it disabled it can process automations on their end before the dice finishing rolling for other users.
 - Bon Mot (Critical Success, Success, Critical Failure)
 - Demoralize (Critical Success, Success)
+- Stunning Blows (Failure, Critical Failure)
 
 ### Changed